changed dir structure add integration test to ci pipeline
