+++ SitNBeer/src/main/java/com/example/demo/SitnbeerApplication.java
@@ -1,0 +1,14 @@
+package com.example.demo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+
+@SpringBootApplication(exclude = {SecurityAutoConfiguration.class })
+public class SitnbeerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SitnbeerApplication.class, args);
+	}
+
+}
