--- src/main/java/com/example/demo/controller/UserController.java
@@ -1,19 +1,0 @@
-package com.example.demo.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-class UserController{
-
-    @RequestMapping("/login")
-	public String login() {
-		return "Hello login";
-    }
-    
-    @RequestMapping("/register")
-	public String register() {
-		return "Hello register";
-	}
-
-}
