--- src/main/resources/application.properties
@@ -1,11 +1,0 @@
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/sitnbeer
-spring.datasource.username=root
-spring.datasource.password=
-spring.jpa.properties.hibernate.show_sql=true
-spring.jpa.properties.hibernate.use_sql_comments=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.type=trace 
-spring.jpa.show-sql=true
-
-server.port=8181
